Contribute screen part 1

+++ feature-crowdloan-impl/src/main/java/jp/co/soramitsu/feature_crowdloan_impl/data/network/api/ParachainMetadata.kt
@@ -8,7 +8,7 @@ fun mapParachainMetadataRemoteToParachainMetadata(parachainMetadata: ParachainMe
             iconLink = icon,
             name = name,
             description = description,
-            rewardRate = rewardRate,
+            rewardRate = rewardRate.toBigDecimal(),
             website = website,
             token = token
         )
